backend/requirements.txt
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.29.0,<1.0
httpx>=0.27.0,<1.0
python-whois>=0.9.4,<1.0
python-dateutil>=2.9.0,<3.0
cryptography>=42.0.8,<44.0
pyOpenSSL>=24.1.0,<25.0
python-dotenv>=1.0.1,<2.0
pydantic>=2.9.0,<3.0
slowapi>=0.1.9,<1.0
playwright>=1.40.0,<2.0
cachetools>=5.3.3,<6.0
